Reorganize to read all data about posts into data structures.
The Ruby mysql libraries seem to read the whole table when the query
is done so there isn't a problem with doing other queries within
a query loop. The C# libraries move through the query one row at a
time so doing a query while processing a query is not possible.